cache/zcache_test.go
package cache

import (
	"fmt"
	"log"
	"testing"
)

/**
 * @author: zurich
 * @Date: 2024/3/15 22:36
 */
var db = map[string]string{
	"zurich": "100",
	"alfred": "92",
	"tomcat": "79",
}

func TestGet(t *testing.T) {
	loadCounts := make(map[string]int, len(db))
	zcache := NewGroup("scores", 2<<10, GetterFunc(
		func(key string) ([]byte, error) {
			log.Println("[SlowDB] search key", key)
			if v, ok := db[key]; ok {
				if _, ok := loadCounts[key]; !ok {
					loadCounts[key] = 0
				}
				loadCounts[key] += 1
				return []byte(v), nil
			}
			return nil, fmt.Errorf("%s not exist", key)
		}))

	for k, v := range db {
		if view, err := zcache.Get(k); err != nil || view.String() != v {
			t.Fatal("failed to get value of Tom")
		} // load from callback function
		if _, err := zcache.Get(k); err != nil || loadCounts[k] > 1 {
			t.Fatalf("cache %s miss", k)
		} // cache hit
	}

	if view, err := zcache.Get("unknown"); err == nil {
		t.Fatalf("the value of unknow should be empty, but %s got", view)
	}
}
